+++ grafanalib/core.py
@@ -80,6 +80,7 @@ GRAPH_TYPE = 'graph'
 DISCRETE_TYPE = 'natel-discrete-panel'
 STAT_TYPE = 'stat'
 SINGLESTAT_TYPE = 'singlestat'
+STATE_TIMELINE_TYPE = 'state-timeline'
 TABLE_TYPE = 'table'
 TEXT_TYPE = 'text'
 ALERTLIST_TYPE = 'alertlist'
@@ -1267,9 +1268,10 @@ class RowPanel(Panel):
     Generates Row panel json structure.
 
     :param title: title of the panel
-    :param panels: list of panels in the row
+    :param collapsed: set True if row should be collapsed
+    :param panels: list of panels in the row, only to be used when collapsed=True
     """
-
+    collapsed = attr.ib(default=False, validator=instance_of(bool))
     panels = attr.ib(default=attr.Factory(list), validator=instance_of(list))
     collapsed = attr.ib(default=False, validator=instance_of(bool))
 
@@ -1941,6 +1943,7 @@ class Stat(Panel):
     :param decimals: number of decimals to display
     :param format: defines value units
     :param graphMode: defines if Grafana will draw graph: keys 'area' 'none'
+    :param noValue: define the default value if no value is found
     :param mappings: the list of values to text mappings
         This should be a list of StatMapping objects
         https://grafana.com/docs/grafana/latest/panels/field-configuration-options/#value-mapping
@@ -1957,6 +1960,7 @@ class Stat(Panel):
     format = attr.ib(default='none')
     graphMode = attr.ib(default='area')
     mappings = attr.ib(default=attr.Factory(list))
+    noValue = attr.ib(default='none')
     orientation = attr.ib(default='auto')
     reduceCalc = attr.ib(default='mean', type=str)
     textMode = attr.ib(default='auto')
@@ -1971,10 +1975,11 @@ class Stat(Panel):
                         'decimals': self.decimals,
                         'mappings': self.mappings,
                         'thresholds': {
-                            'mode': 'absolute',
+                            'mode': ABSOLUTE_TYPE,
                             'steps': self.thresholds,
                         },
-                        'unit': self.format
+                        'unit': self.format,
+                        'noValue': self.noValue
                     }
                 },
                 'options': {
@@ -2417,109 +2422,70 @@ class Column(object):
         }
 
 
-def _style_columns(columns):
-    """Generate a list of column styles given some styled columns.
-
-    The 'Table' object in Grafana separates column definitions from column
-    style definitions. However, when defining dashboards it can be very useful
-    to define the style next to the column. This function helps that happen.
-
-    :param columns: A list of (Column, ColumnStyle) pairs. The associated
-        ColumnStyles must not have a 'pattern' specified. You can also provide
-       'None' if you want to use the default styles.
-    :return: A list of ColumnStyle values that can be used in a Grafana
-        definition.
-    """
-    new_columns = []
-    styles = []
-    for column, style in columns:
-        new_columns.append(column)
-        if not style:
-            continue
-        if style.pattern and style.pattern != column.text:
-            raise ValueError(
-                "ColumnStyle pattern (%r) must match the column name (%r) if "
-                "specified" % (style.pattern, column.text))
-        styles.append(attr.evolve(style, pattern=column.text))
-    return new_columns, styles
-
-
 @attr.s
 class Table(Panel):
     """Generates Table panel json structure
 
-    Grafana doc on table: https://grafana.com/docs/grafana/latest/features/panels/table_panel/#table-panel
+    Now supports Grafana v8+
+    Grafana doc on table: https://grafana.com/docs/grafana/latest/visualizations/table/
 
-    :param columns: table columns for Aggregations view
-    :param fontSize: defines value font size
-    :param pageSize: rows per page (None is unlimited)
-    :param scroll: scroll the table instead of displaying in full
-    :param showHeader: show the table header
-    :param sort: table sorting
-    :param styles: defines formatting for each column
-    :param transform: table style
+    :param align: Align cell contents; auto (default), left, center, right
+    :param colorMode: Default thresholds
+    :param columns: Table columns for Aggregations view
+    :param displayMode: By default, Grafana automatically chooses display settings, you can choose;
+        color-text, color-background, color-background-solid, gradient-gauge, lcd-gauge, basic, json-view
+    :param fontSize: Defines value font size
+    :param filterable: Allow user to filter columns, default False
+    :param mappings: To assign colors to boolean or string values, use Value mappings
+    :param showHeader: Show the table header
+    :param thresholds: List of thresholds
     """
 
+    align = attr.ib(default='auto', validator=instance_of(str))
+    colorMode = attr.ib(default='thresholds', validator=instance_of(str))
     columns = attr.ib(default=attr.Factory(list))
+    displayMode = attr.ib(default='auto', validator=instance_of(str))
     fontSize = attr.ib(default='100%')
-    pageSize = attr.ib(default=None)
-    scroll = attr.ib(default=True, validator=instance_of(bool))
+    filterable = attr.ib(default=False, validator=instance_of(bool))
+    mappings = attr.ib(default=attr.Factory(list))
     showHeader = attr.ib(default=True, validator=instance_of(bool))
     span = attr.ib(default=6)
-    sort = attr.ib(
-        default=attr.Factory(ColumnSort), validator=instance_of(ColumnSort))
-    styles = attr.ib()
-
-    transform = attr.ib(default=COLUMNS_TRANSFORM)
-
-    @styles.default
-    def styles_default(self):
-        return [
-            ColumnStyle(
-                alias='Time',
-                pattern='Time',
-                type=DateColumnStyleType(),
-            ),
-            ColumnStyle(
-                alias='time',
-                pattern='time',
-                type=DateColumnStyleType(),
-            ),
-            ColumnStyle(
-                pattern='/.*/',
-            ),
-        ]
+    thresholds = attr.ib(default=attr.Factory(list))
 
     @classmethod
     def with_styled_columns(cls, columns, styles=None, **kwargs):
-        """Construct a table where each column has an associated style.
-
-        :param columns: A list of (Column, ColumnStyle) pairs, where the
-            ColumnStyle is the style for the column and does not have a
-            pattern set (or the pattern is set to exactly the column name).
-            The ColumnStyle may also be None.
-        :param styles: An optional list of extra column styles that will be
-            appended to the table's list of styles.
-        :param kwargs: Other parameters to the Table constructor.
-        :return: A Table.
-        """
-        extraStyles = styles if styles else []
-        columns, styles = _style_columns(columns)
-        return cls(columns=columns, styles=styles + extraStyles, **kwargs)
+        """Styled columns is not support in Grafana v8 Table"""
+        print("Error: Styled columns is not support in Grafana v8 Table")
+        print("Please see https://grafana.com/docs/grafana/latest/visualizations/table/ for more options")
+        raise NotImplementedError
 
     def to_json_data(self):
         return self.panel_json(
             {
+                "color": {
+                    "mode": self.colorMode
+                },
                 'columns': self.columns,
                 'fontSize': self.fontSize,
+                'fieldConfig': {
+                    'defaults': {
+                        'custom': {
+                            'align': self.align,
+                            'displayMode': self.displayMode,
+                            'filterable': self.filterable
+                        },
+                        "thresholds": {
+                            "mode": "absolute",
+                            "steps": self.thresholds
+                        }
+                    }
+                },
                 'hideTimeOverride': self.hideTimeOverride,
+                'mappings': self.mappings,
                 'minSpan': self.minSpan,
-                'pageSize': self.pageSize,
-                'scroll': self.scroll,
-                'showHeader': self.showHeader,
-                'sort': self.sort,
-                'styles': self.styles,
-                'transform': self.transform,
+                'options': {
+                    'showHeader': self.showHeader
+                },
                 'type': TABLE_TYPE,
             }
         )
@@ -3283,5 +3249,74 @@ class Worldmap(Panel):
                     'metricField': 'metric'
                 },
                 'type': WORLD_MAP_TYPE
+            }
+        )
+
+
+@attr.s
+class StateTimeline(Panel):
+    """Generates State Timeline panel json structure
+    Grafana docs on State Timeline panel: https://grafana.com/docs/grafana/latest/visualizations/state-timeline/
+
+    :param alignValue: Controls value alignment inside state regions, default left
+    :param colorMode: Default thresholds
+    :param fillOpacity: Controls the opacity of state regions, default 0.9
+    :param legendDisplayMode: refine how the legend appears, list, table or hidden
+    :param legendPlacement: bottom or top
+    :param lineWidth: Controls line width of state regions
+    :param mappings: To assign colors to boolean or string values, use Value mappings
+    :param mergeValues: Controls whether Grafana merges identical values if they are next to each other, default True
+    :param rowHeight: Controls how much space between rows there are. 1 = no space = 0.5 = 50% space
+    :param showValue: Controls whether values are rendered inside the state regions. Auto will render values if there is sufficient space.
+    :param tooltipMode: Default single
+    :param thresholds: Thresholds are used to turn the time series into discrete colored state regions
+    """
+    alignValue = attr.ib(default='left', validator=instance_of(str))
+    colorMode = attr.ib(default='thresholds', validator=instance_of(str))
+    fillOpacity = attr.ib(default=70, validator=instance_of(int))
+    legendDisplayMode = attr.ib(default='list', validator=instance_of(str))
+    legendPlacement = attr.ib(default='bottom', validator=instance_of(str))
+    lineWidth = attr.ib(default=0, validator=instance_of(int))
+    mappings = attr.ib(default=attr.Factory(list))
+    mergeValues = attr.ib(default=True, validator=instance_of(bool))
+    rowHeight = attr.ib(default=0.9, validator=instance_of(float))
+    showValue = attr.ib(default='auto', validator=instance_of(str))
+    tooltipMode = attr.ib(default='single', validator=instance_of(str))
+    thresholds = attr.ib(default=attr.Factory(list))
+
+    def to_json_data(self):
+        return self.panel_json(
+            {
+                'fieldConfig': {
+                    'defaults': {
+                        'custom': {
+                            'lineWidth': self.lineWidth,
+                            'fillOpacity': self.fillOpacity
+                        },
+                        'color': {
+                            'mode': self.colorMode
+                        },
+                        'thresholds': {
+                            'mode': ABSOLUTE_TYPE,
+                            'steps': self.thresholds,
+                        },
+                        'mappings': self.mappings
+                    },
+                    'overrides': []
+                },
+                'options': {
+                    'mergeValues': self.mergeValues,
+                    'showValue': self.showValue,
+                    'alignValue': self.alignValue,
+                    'rowHeight': self.rowHeight,
+                    'legend': {
+                        'displayMode': self.legendDisplayMode,
+                        'placement': self.legendPlacement
+                    },
+                    'tooltip': {
+                        'mode': self.tooltipMode
+                    }
+                },
+                'type': STATE_TIMELINE_TYPE,
             }
         )